pyproject.toml
[tool.poetry]
name = "ai-cms"
version = "0.1.0"
description = ""
authors = ["Piyush289kumar <piyushraikwar289@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django-unfold = "^0.55.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
